Move shouldIncludeScientificReviewNoticeWhenLevelDisplayX to FetchSerialPublicationByIdentifierAndYearHandlerTest
